Moved utility method to IOUtils
