+++ solr/core/src/java/org/apache/solr/handler/loader/XMLLoader.java
@@ -46,6 +46,7 @@ import org.apache.solr.common.params.UpdateParams;
 import org.apache.solr.common.util.CollectionUtil;
 import org.apache.solr.common.util.ContentStream;
 import org.apache.solr.common.util.ContentStreamBase;
+import org.apache.solr.common.util.IOUtils;
 import org.apache.solr.common.util.StrUtils;
 import org.apache.solr.common.util.XMLErrorLogger;
 import org.apache.solr.handler.RequestHandlerUtils;
@@ -137,9 +138,7 @@ public class XMLLoader extends ContentStreamLoader {
               "body: {}",
               new String(
                   body,
-                  (charset == null)
-                      ? StandardCharsets.UTF_8
-                      : ContentStreamBase.charsetForName(charset)));
+                  (charset == null) ? StandardCharsets.UTF_8 : IOUtils.charsetForName(charset)));
         }
         return new ByteArrayInputStream(body);
       }
