+++ solr/core/src/java/org/apache/solr/spelling/FileBasedSpellChecker.java
@@ -33,7 +33,7 @@ import org.apache.lucene.search.spell.HighFrequencyDictionary;
 import org.apache.lucene.search.spell.PlainTextDictionary;
 import org.apache.lucene.store.ByteBuffersDirectory;
 import org.apache.lucene.store.Directory;
-import org.apache.solr.common.util.ContentStreamBase;
+import org.apache.solr.common.util.IOUtils;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.core.SolrCore;
 import org.apache.solr.schema.FieldType;
@@ -124,7 +124,7 @@ public class FileBasedSpellChecker extends AbstractLuceneSpellChecker {
               new PlainTextDictionary(
                   new InputStreamReader(
                       core.getResourceLoader().openResource(sourceLocation),
-                      ContentStreamBase.charsetForName(characterEncoding)));
+                      IOUtils.charsetForName(characterEncoding)));
         }
       }
