+++ solr/solrj/src/java/org/apache/solr/common/util/IOUtils.java
@@ -16,7 +16,9 @@
  */
 package org.apache.solr.common.util;
 
+import java.io.UnsupportedEncodingException;
 import java.lang.invoke.MethodHandles;
+import java.nio.charset.Charset;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,6 +37,23 @@ public class IOUtils {
       }
     } catch (Exception e) {
       log.error("Error while closing", e);
+    }
+  }
+
+  /**
+   * Resolves a charset name (typically from a Content-Type header) to a {@link Charset}. Unlike
+   * {@link Charset#forName(String)}, an illegal or unsupported name results in a checked {@link
+   * UnsupportedEncodingException}, matching the behavior of the legacy {@code String}-based JDK
+   * charset APIs, so callers can treat a bad charset as an I/O error.
+   */
+  public static Charset charsetForName(String charsetName) throws UnsupportedEncodingException {
+    try {
+      return Charset.forName(charsetName);
+    } catch (IllegalArgumentException e) {
+      UnsupportedEncodingException uee =
+          new UnsupportedEncodingException("Unsupported charset: " + charsetName);
+      uee.initCause(e);
+      throw uee;
     }
   }
 }
